% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.event.R
\name{basestats}
\alias{basestats}
\alias{Constant}
\alias{CovSnd}
\alias{CovRec}
\alias{CovEvent}
\alias{CovInt}
\title{Format event statistics of different types}
\usage{
Constant(data)

CovSnd(stat.func = NULL, stats = NULL, data = NULL,
  time.varying = TRUE, ...)

CovRec(stat.func = NULL, stats = NULL, data = NULL,
  time.varying = TRUE, ...)

CovEvent(stat.func = NULL, stats = NULL, data = NULL,
  time.varying = TRUE, ...)

CovInt(stat.func = NULL, stats = NULL, data = NULL,
  time.varying = TRUE, ...)
}
\arguments{
\item{data}{......}

\item{stat.func}{A function used to compute statistics}

\item{stats}{An array containing pre-computed statistics}

\item{time.varying}{Boolean value indicating whether statistic should be
computed at each time point}

\item{...}{......}
}
\value{
An array of statistics of dimension (S x R x E).
}
\description{
These functions sweep out statistics to construct constant effects,
sender effects, receiver effects, dyadic effects, and unordered pair effects.
}
