% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.stats.R
\name{combine.stats}
\alias{combine.stats}
\title{Format Event Network Statistics}
\usage{
combine.stats(...)
}
\arguments{
\item{...}{Named arguments with statistics}
}
\value{
A 4 dimensional array of statistics
}
\description{
Takes network statistics as arguments and combines them into a
4 dimensional array
}
