% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{constraints}
\alias{constraints}
\title{CFA Constraints Matrix}
\usage{
constraints(X, order)
}
\arguments{
\item{X}{numeric data frame containing all the variables needed in the factor analysis}

\item{order}{original order of variables in the data frame}
}
\value{
matrix of zeros and ones indicating the relationship that wants to be modeled
}
\description{
Function to create constraints matrix based on user inputs for Confirmatory Factor Analysis
}
\references{
Nieser, K. J., & Cochran, A. L. (2021). Addressing heterogeneous populations in latent variable settings through robust estimation. Psychological Methods.
}
\author{
Bryan Ortiz-Torres (bortiztorres@wisc.edu); Kenneth Nieser (nieser@wisc.edu)
}
