% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controlREM.R
\name{controlREM}
\alias{controlREM}
\title{Control parameters for REM package}
\usage{
controlREM(
  steps = 25,
  tol = 1e-06,
  maxiter = 1000,
  min_weights = 1e-30,
  max_ueps = 0.3,
  chk_gamma = 0.9,
  n = 20000
)
}
\arguments{
\item{steps}{maximum number of epsilons considered in the analysis (default = 25).}

\item{tol}{tolerance parameter to check for convergence (default = 1e-6).}

\item{maxiter}{maximum iterations allowed to calculate the estimates (default = 1e3)}

\item{min_weights}{Lower bound for the individual weights estimated by REM}

\item{max_ueps}{The largest epsilon value to check when searching for the optimal epsilon.}

\item{chk_gamma}{gamma value used when searching for epsilon}

\item{n}{The simulated n in checkEpsFA}
}
\value{
control parameters used in the REM package (steps, tol, maxiter, min_weights,ueps,n).
}
\description{
Control parameters for REM package
}
\references{
Nieser, K. J., & Cochran, A. L. (2021). Addressing heterogeneous populations in latent variable settings through robust estimation. Psychological Methods.
}
\seealso{
\code{\link[=REM_EFA]{REM_EFA()}}, \code{\link[=REM_CFA]{REM_CFA()}}
}
\author{
Bryan Ortiz-Torres (bortiztorres@wisc.edu); Kenneth Nieser (nieser@wisc.edu)
}
