% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.degree.R
\name{degreestats}
\alias{degreestats}
\alias{IDSnd}
\alias{IDRec}
\alias{ODSnd}
\alias{ODRec}
\alias{TDSnd}
\alias{TDRec}
\alias{NIDSnd}
\alias{NIDRec}
\alias{NODSnd}
\alias{NODRec}
\alias{NTDSnd}
\alias{NTDRec}
\title{Degree statistics}
\usage{
IDSnd(data)

IDRec(data)

ODSnd(data)

ODRec(data)

TDSnd(data)

TDRec(data)

NIDSnd(data)

NIDRec(data)

NODSnd(data)

NODRec(data)

NTDSnd(data)

NTDRec(data)
}
\arguments{
\item{data}{Array (E x 3) of events}
}
\value{
Array of E x S x R
}
\description{
Compute in/out/total degree sender and receiver effects, normalized or unnormalized
}
