% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.event.network.core.R
\name{FitEventNetworkCore}
\alias{FitEventNetworkCore}
\title{Core Relational Event Model fitting function}
\usage{
FitEventNetworkCore(sequence, stats, x0 = rep(0, dim(stats)[3]),
  ordinal = FALSE, blocked = c(), epsilon = 1e-06, iter_max = 200,
  networkStatTerms = 1:dim(stats)[3], ...)
}
\arguments{
\item{sequence}{An (E x 3) matrix of values where the first column is the event
sender, the second column is the event receiver, and the last column is the
time}

\item{stats}{An array of dimension (S x R x K x E). An entry (i,j,k,t) is
interpreted as the value of statistic k
from sender i to receiver j at event time t}

\item{x0}{a vector of length numStats - User-supplied initial point. We can
possibly make it so there is a default here}

\item{ordinal}{Binary variable, where TRUE = ordinal likelihood (order
only) and FALSE = temporal (full time info).}

\item{blocked}{(not implemented) a list of indices for the events that cannot happen. As of
now, this is implemented as just numbers corresponding to matrix locations.}

\item{epsilon}{Convergence criterion.}

\item{iter_max}{Iteration limit}

\item{networkStatTerms}{A numerical list of the terms to be included in the
model. These numbers should correspond to the statistic from the list in the
provided array.}

\item{...}{more functions}
}
\value{
An object of class 'rem.fit'
}
\description{
Core Relational Event Model fitting function
}
