% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.fixed.R
\name{fixedstats}
\alias{fixedstats}
\alias{FESnd}
\alias{FERec}
\alias{FEInt}
\title{Generates fixed effects}
\usage{
FESnd(data)

FERec(data)

FEInt(data)
}
\arguments{
\item{data}{...}
}
\value{
Array of E x S x R
}
\description{
Fixed sender/receiver/unordered pair effects
}
