% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.fraction.R
\name{fractionstats}
\alias{fractionstats}
\alias{FrSndSnd}
\alias{FrRecSnd}
\title{Fraction of event statistics}
\usage{
FrSndSnd(data)

FrRecSnd(data)
}
\arguments{
\item{data}{Array (E x 3) of events}
}
\value{
Array of E x S x R
}
\description{
Compute effects of fraction of sending or receiving effects on sending
}
