% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{gen_Y}
\alias{gen_Y}
\title{Outcome Generation Function}
\usage{
gen_Y(X, beta, tau = 1, mu = 0, Sigma_e = 1, seed = NULL)
}
\arguments{
\item{X}{Numeric matrix; n × K covariate matrix.}

\item{beta}{Numeric vector; coefficient vector of length K.}

\item{tau}{Numercalar; average treatment effect. Default is 1.}

\item{mu}{Numeric scalar; intercept term. Default is 0.}

\item{Sigma_e}{Numeric scalar; standard deviation of the error term. Default is 1.}

\item{seed}{Integer scalar; random seed for reproducibility. Default is NULL (no seed set).}
}
\value{
A list containing:
\describe{
\item{Y0}{Numeric vector; outcomes under control with length n.}
\item{Y1}{Numeric vector; outcomes under treatment with length n.}
}
}
\description{
Generates potential outcomes Y(0) and Y(1) based on a linear model with covariates X, coefficients beta, treatment effect tau, and error term. If a treatment assignment vector Z is provided, it also computes the observed outcomes Y.
}
\examples{
set.seed(123)
X <- gen_X(n = 100, K = 5)
beta <- gen_beta(Sigma_X = diag(5), R2 = 0.5)
Y <- gen_Y(X = X, beta = beta)

}
