\name{get.triad.stat}
\alias{get.triad.stat}
\alias{triad}
\alias{path-dependency}
\title{Calculate triad statistics}
%
\description{Calculate the endogenous network statistic \empth{inertia} that 
measures the tendency for events to consist of the same sender and target 
(i.e. repeated events).}
\usage{
get.triad.stat(data, time, sender, target, halflife, 
  weight = NULL, eventtypevar = NULL, eventtypevalues = NULL,
  eventattributevar = NULL, eventattributeAI = NULL,
  eventattributeBI = NULL, eventattributeAB = NULL,
  variablename = "triad", returnData = FALSE, 
  showprogressbar = FALSE)
}
\arguments{
\item{data}{}
\item{time}{}
\item{sender}{}
\item{target}{}
\item{halflife}{}
\item{weight}{}
\item{eventtypevar}{}
\item{eventtypevalues}{}
\item{eventattributevar}{}
\item{eventattributeAI}{}
\item{eventattributeBI}{}
\item{eventattributeAB}{}
\item{variablename}{}
\item{returnData}{}
\item{showprogressbar}{}
}
\details{

}
\value{

}
\references{

}
\note{

}
\author{
Laurence Brandenberger \email{laurence.brandenberger@eawag.ch}
}
\seealso{
\code{\link{rem-package}} 
}
\examples{

}
\keyword{key}




