% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.group.R
\name{groupstats}
\alias{groupstats}
\alias{SameConstGroup}
\alias{DiffConstGroup}
\title{Group statistics}
\usage{
SameConstGroup(data, covar)

DiffConstGroup(data, covar)
}
\arguments{
\item{data}{data file}

\item{covar}{covariates}
}
\description{
Group statistics
}
