% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.participation.R
\name{pshiftstats}
\alias{pshiftstats}
\alias{PSAB.BA}
\alias{PSAB.BY}
\alias{PSAB.XA}
\alias{PSAB.XB}
\alias{PSAB.XY}
\alias{PSAB.AY}
\title{Participation shift event statistics}
\usage{
PSAB.BA(data)

PSAB.BY(data)

PSAB.XA(data)

PSAB.XB(data)

PSAB.XY(data)

PSAB.AY(data)
}
\arguments{
\item{data}{Array (E x 3) of events}
}
\value{
Array of E x S x R
}
\description{
Compute pshift effects
}
