% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.recency.R
\name{recencystats}
\alias{recencystats}
\alias{RRecSnd}
\alias{RSndSnd}
\alias{RERecSnd}
\alias{RESndSnd}
\title{Recency statistics}
\usage{
RRecSnd(data)

RSndSnd(data)

RERecSnd(data)

RESndSnd(data)
}
\arguments{
\item{data}{Array (E x 3) of events}
}
\value{
Array of E x S x R
}
\description{
Compute effects of recency of receipt or sending on sending or recieving
in terms of time or number of events. Name format:
\code{R}[\code{E}vent][recency of \code{rec}/\code{snd}][effect on \code{snd}/\code{rec}]
}
