\name{rem-package}
\alias{REM}
\alias{relational-event-model}
\alias{relational event model}
\alias{rem-package}
\alias{rem}
\docType{package}
\title{Fit Relational Event Models (REM)}
\description{
rem can be used to fit relational event models 
(REM). REMs are used to check for network dependencies among events embedded in 
an event sequence. The package includes functions to calculate (signed) 
endogenous network statistics to test for network dependencies among events
embedded in an event sequence. A large number of one-mode/two-mode/multi-mode 
network  statistics can be calculated, such as inertia, reciprocity, triadic 
closure (triads), closing four-cycles, absolute differences or 
similarity-measures.
}
\details{
\tabular{ll}{
Package: \tab rem\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-08-17\cr
License: \tab What license is it under?\cr
}
- relational event models
- flexible time decay function: more recent = more important
- rate and type estimation
- test whether event sequence is important: distinguish between homophily/influence
- calculate endogenous network statistics
- plot statistics
- gof (tbd)

~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Laurence Brandenberger <laurence.brandenberger@eawag.ch>
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
