% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rem.stat.R
\name{rem.stat}
\alias{rem.stat}
\title{Sweep out network statistics}
\usage{
rem.stat(stat.func = NULL, stats = NULL, data, time.varying = TRUE,
  dyadic = TRUE, sender = TRUE, receiver = FALSE, either = FALSE,
  effects = "sendingrate", normalize = FALSE)
}
\arguments{
\item{stat.func}{Statistic function \code{\link{remstats}}}

\item{stats}{Raw data array of appropriate dimension: E x S, E x R, S x R, or E x S x R}

\item{data}{Event data (E x 3)}

\item{time.varying}{Logical}

\item{dyadic}{Logical}

\item{sender}{Logical}

\item{receiver}{Logical}

\item{either}{Logical value indicating whether order matters}

\item{effects}{Character indicates what the statistic effects ('sendingrate' or 'recvingrate').
Transposes array as necessary}

\item{normalize}{Logical - indicating whether should be zero centered with sd of 1}
}
\value{
Array of size E x S x R
}
\description{
Basic function to calculate network statistics and format proper statistic arrays.
Compatible with both functions and raw data.
}
