% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REM_CFA.R
\name{REM_CFA}
\alias{REM_CFA}
\title{Robust Estimation Maximization Estimates for Confirmatory Factor Analysis}
\usage{
REM_CFA(X, delta = 0.05, model = NA, ctrREM = controlREM())
}
\arguments{
\item{X}{numeric data frame containing all the variables needed in the factor analysis}

\item{delta}{hyperparameter between 0 and 1 that captures the researcher’s tolerance of incorrectly down-weighting data from the model (default = 0.05).}

\item{model}{string variable that contains each structural equation in a new line where equalities are denoted by the symbol "~".}

\item{ctrREM}{provides control parameters (default: controlREM(steps = 25, tol = 1e-6, maxiter = 1e3))}
}
\value{
REM_CFA returns an object of class "REM". The function \code{\link[=summary]{summary()}} is used to obtain estimated parameters from the model. An object of class "REM" in Confirmatory Factor Analysis is a list of outputs with four different components: the matched call (call), estimates using traditional expectation maximization (EM_output), estimates using robust expectation maximization (REM_output), and a summary table (summary_table). The list contains the following components:
\item{call}{match call}
\item{model}{model frame}
\item{delta}{hyperparameter between 0 and 1 that captures the researcher’s tolerance of incorrectly down-weighting data from the model}
\item{k}{number of factors}
\item{constraints}{p x k matrix of zeros and ones denoting the factors (rows) and observed variables (columns)}
\item{epsilon}{hyperparameter on the likelihood scale}
\item{AIC_rem}{Akaike Information Criterion}
\item{BIC_rem}{Bayesian Information Criterion}
\item{mu}{intercept}
\item{lambda}{loadings}
\item{psi}{unique variance}
\item{gamma}{average weights}
\item{weights}{weights}
\item{ind_lik}{likelihood value for each individual}
\item{lik_rem}{joint log-likelihood evaluated at REM estimates}
\item{lik}{joint log-likelihood evaluated at EM estimates}
\item{method}{indicator for CFA or EFA}
\item{parameter}{parameters presented}
\item{estimates}{estimates from the analisis}
\item{se}{standard errors}
\item{Z}{z statistic}
\item{pvalue}{p-value}
\item{ci.lower}{lower 95\% confidence interval}
\item{ci.upper}{upper 95\% confidence interval}
}
\description{
Robust Estimation Maximization Estimates for Confirmatory Factor Analysis
}
\references{
Nieser, K. J., & Cochran, A. L. (2021). Addressing heterogeneous populations in latent variable settings through robust estimation. Psychological Methods.
}
\seealso{
\code{\link[=REM_EFA]{REM_EFA()}}, \code{\link[=summary]{summary()}}
}
\author{
Bryan Ortiz-Torres (bortiztorres@wisc.edu); Kenneth Nieser (nieser@wisc.edu)
}
