% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REM_EFA.R
\name{REM_EFA}
\alias{REM_EFA}
\title{Robust Estimation Maximization for Exploratory Factor Analysis}
\usage{
REM_EFA(X, k_range, delta = 0.05, rotation = "oblimin", ctrREM = controlREM())
}
\arguments{
\item{X}{numeric data frame containing all the variables needed in the factor analysis}

\item{k_range}{vector with the range of number of factors you want to look at}

\item{delta}{hyperparameter between 0 and 1 that captures the researcher’s tolerance of incorrectly down-weighting data from the model (default = 0.05).}

\item{rotation}{factor rotation method (either 'oblimin' or 'varimax')}

\item{ctrREM}{provides control parameters (default: controlREM(steps = 25, tol = 1e-6, maxiter = 1e3))}
}
\value{
REM_CFA returns an object of class "REM". The function \code{\link[=summary]{summary()}} is used to obtain estimated parameters from the model. An object of class "REM" in Exploratory Factor Analysis is a list of outputs with four different components for each number of factor: the matched call (call), estimates using traditional expectation maximization (EM_output), estimates using robust expectation maximization (REM_output), and a summary table (summary_table). The list contains the following components:
\item{call}{match call}
\item{model}{model frame}
\item{k}{number of factors}
\item{constraints}{p x k matrix of zeros and ones denoting the factors (rows) and observed variables (columns)}
\item{epsilon}{hyperparameter on the likelihood scale}
\item{AIC_rem}{Akaike Information Criterion}
\item{BIC_rem}{Bayesian Information Criterion}
\item{mu}{intercept}
\item{lambda}{loadings}
\item{psi}{unique variance}
\item{phi}{factor covariance matrix}
\item{gamma}{average weights}
\item{weights}{weights}
\item{ind_lik}{likelihood value for each individual}
\item{lik_rem}{joint log-likelihood evaluated at REM estimates}
\item{lik}{joint log-likelihood evaluated at EM estimates}
\item{mu.se}{standard error of the intercept}
\item{lambda.se}{standard error of the loadings}
\item{psi.se}{standard error of the unique variance}
\item{gamma.se}{standard error of the average weights}
\item{method}{indicator for CFA or EFA}
\item{parameter}{parameters presented}
\item{estimates}{estimates from the analisis}
\item{se}{standard errors}
\item{Z}{z statistic}
\item{pvalue}{p-value}
\item{ci.lower}{lower 95\% confidence interval}
\item{ci.upper}{upper 95\% confidence interval}

The summary function can be used to obtain estimated parameters from the optimal model based on the BIC from the EM and REM algorithms.
}
\description{
The robust expectation maximization (REM) algorithm is a statistical method for estimating the parameters of a latent variable model in the presence of population heterogeneity as suggested by Nieser & Cochran (2021).
The REM algorithm is based on the expectation-maximization (EM) algorithm, but it allows for the case when not all the data are generated by the assumed data generating model.
}
\references{
Nieser, K. J., & Cochran, A. L. (2021). Addressing heterogeneous populations in latent variable settings through robust estimation. Psychological Methods.
}
\seealso{
\code{\link[=summary.REM]{summary.REM()}} for more detailed summaries, \code{\link[=oblimin]{oblimin()}} and \code{\link[=varimax]{varimax()}} for details on the rotation
}
\author{
Bryan Ortiz-Torres (bortiztorres@wisc.edu); Kenneth Nieser (nieser@wisc.edu)
}
