% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitive.functions.R
\name{remstats}
\alias{remstats}
\alias{time_since}
\alias{events_since}
\alias{event.count}
\alias{fixed.effect}
\alias{alter.count}
\alias{outbound.path.count}
\alias{inbound.path.count}
\alias{outbound.shared.partner.count}
\alias{inbound.shared.partner.count}
\alias{pshift}
\title{Primitive functions for computing statistics.}
\usage{
time_since(X, s = NA, r = NA, t)

events_since(X, s = NA, r = NA, t)

event.count(X, s = NA, r = NA, t)

fixed.effect(X, s = NA, r = NA, t)

alter.count(X, s, r, t, sndtype = "rec", rectype = "snd")

outbound.path.count(X, s, r, t)

inbound.path.count(X, s, r, t)

outbound.shared.partner.count(X, s, r, t)

inbound.shared.partner.count(X, s, r, t)

pshift(X, s, r, t, s2 = NA, r2 = NA, ordinal = TRUE)
}
\arguments{
\item{X}{An Ex3 matrix of values where the first column is the event
sender, the second column is the event receiver, and the last column is the
time}

\item{s}{The sender index. Use NA to ignore}

\item{r}{The reciever index. Use NA to ignore}

\item{t}{time}

\item{sndtype}{Type of alter of first node ('rec'/'snd')}

\item{rectype}{Type of alter of second node ('snd'/'rec')}

\item{s2}{Sender index of second dyad. Use NA to ignore.}

\item{r2}{Receiver index of second dyad. Use NA to ignore.}

\item{ordinal}{Logical. Whether recency is calculated in number of events or time. Not implemented}
}
\value{
A scalar statistic for the given sender/receiver/dyad at time t
}
\description{
Each function takes four or more arguments, a dataset, a sender index, a receiver
index, and a time. These are meant to be used in conjunction with sweeping functions
to generate appropriate statistics arrays. These functions are general purpose.
They are intended to work for sender, receiver, dyadic, and unordered pair effects.
\code{time_since} returns the time since event last occurred.
\code{events_since} returns the number of events since event last occurred.
\code{event.count} returns the number of times this event has already occurred.
\code{fixed.effect} returns the indicator for whether event at time t has sender s and receiver r.
\code{alter.count} returns the number of alters of a certain type.
\code{outbound.path.count} returns a count of the number of outbound transitive paths.
\code{inbound.path.count} returns a count of the number of inbound transitive paths.
\code{outbound.shared.partner.count} returns a count of the number of outbound shared partners.
\code{inbound.shared.partner.count} returns a count of the number of inbound shared partners.
\code{pshift} returns the indicator for whether a pshift has occurred.
}
