% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.event.network.fit.R,
%   R/summary.event.network.fit.R
\name{print.summary.event.network.fit}
\alias{print.summary.event.network.fit}
\alias{summary.event.network.fit}
\title{Summarizing Event Network Fits}
\usage{
\method{print}{summary.event.network.fit}(x, ...)

\method{summary}{event.network.fit}(object, ...)
}
\arguments{
\item{x}{object of class `summary.event.network.fit` returned by
[summary.event.network.fit()].}

\item{...}{....}

\item{object}{an object of class \code{"event.network.fit"}}
}
\value{
The function \code{\link{summary.event.network.fit}} computes and
  returns a list of summary statistics of the fitted REM  given in \code{object}.
}
\description{
Summarizing Event Network Fits
}
