% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.triadic.R
\name{triadstats}
\alias{triadstats}
\alias{OTPSnd}
\alias{ITPSnd}
\alias{OSPSnd}
\alias{ISPSnd}
\title{Triadic event statistics}
\usage{
OTPSnd(data)

ITPSnd(data)

OSPSnd(data)

ISPSnd(data)
}
\arguments{
\item{data}{Array (E x 3) of events}
}
\value{
Array of E x S x R
}
\description{
Compute effects of outbound/inbound transitive path/shared path
effects on sending
}
